+++ src/SFA.DAS.ForecastingTool.Web/FinancialForecasting/IForecastCalculator.cs
@@ -8,6 +8,6 @@ namespace SFA.DAS.ForecastingTool.Web.FinancialForecasting
     public interface IForecastCalculator
     {
         
-        Task<ForecastResult> ForecastAsync(int paybill, int englishFraction, StandardModel[] myStandards, int duration);
+        Task<ForecastResult> ForecastAsync(long paybill, int englishFraction, StandardModel[] myStandards, int duration);
     }
 }
